Controller 이전 코드

+++ src/main/kotlin/main.kt
@@ -7,10 +7,18 @@ fun readLineTrim() = readLine()!!.trim()
 fun main() {
     println("== SIMPLE SSG 시작 ==")
 
+    memberRepository.makeTestMembers()
     articleRepository.makeTestArticle()
 
+    var loginedMember: Member? = null //로그인 상태 식별
+
     while (true) {
-        print("명령어> ")
+        val propmpt = if (loginedMember == null){
+            "명령어> "
+        } else {
+            "${loginedMember.nickname}> "
+        }
+        print(propmpt)
         val command = readLineTrim()
         
         val rq = Rq(command)
@@ -20,11 +28,87 @@ fun main() {
                 println("시스템 종료")
                 break
             }
+
+            "/member/logout" -> {
+                loginedMember = null
+                println("로그아웃.")
+            }
+
+            "/member/login" -> {
+                print("로그인아이디 : ")
+                val loginId = readLineTrim()
+
+                val member = memberRepository.getMemberByLoginedId(loginId) // 현재 저장되어있는 아이디인지 확인
+                if (member == null) {
+                    println("'$loginId'(은)는 존재하지 않은 아이디입니다")
+                    continue
+                }
+                print("로그인비밀번호 : ")
+                val loginPw = readLineTrim()
+                if (member.loginPw != loginPw) {
+                    print("비밀번호가 일치하지 않습니다")
+                    continue
+                }
+                loginedMember = member // 로그인 된 멤버의 객체를 넘겨 로그인 상태 유지 체크
+
+                println("'${member.nickname}'님 환영합니다.")
+            }
+
+            // 새로운 멤버 추가
+            "/member/join" -> {
+                print("로그인아이디 : ")
+                val loginId = readLineTrim()
+
+                // 파라메터로 loginId를 넘겨 멤버가 이미 존재하는지 확인
+                val isJoinableLoginId = memberRepository.isJoinableLoginId(loginId)
+                if (isJoinableLoginId == false) {
+                    println("'$loginId'(은)는 이미 사용중인 아이디입니다")
+                    continue
+                }
+
+                print("로그인비밀번호 : ")
+                val loginPw = readLineTrim()
+                print("이름 : ")
+                val name = readLineTrim()
+                print("별명 : ")
+                val nickname = readLineTrim()
+                print("휴대전화번호 : ")
+                val cellphoneNo = readLineTrim()
+                print("이메일 : ")
+                val email = readLineTrim()
+
+                val id = memberRepository.join(loginId, loginPw, name, nickname, cellphoneNo, email)
+
+                // 이전에 id가 9이면 새로운 가입자의 id는 '10'이 됨.
+                println("${id}번 회원으로 가입되었습니다.")
+            }
+
+            "/article/write" -> {
+                if (loginedMember == null) {
+                    println("로그인해주세요.")
+                    continue
+                }
+
+                print("제목 : ")
+                val title = readLineTrim()
+                print("내용 : ")
+                val body = readLineTrim()
+
+                val id = articleRepository.addArticle(loginedMember.id, title, body)
+
+                println("${id}번 게시물이 추가되었습니다.")
+            }
+
             "/article/detail" -> {
+                if (loginedMember == null) {
+                    println("로그인해주세요.")
+                    continue
+                }
+
                 val id = rq.getIntParam("id", 0)
 
                 if (id == 0){
-                    println("id를 입력해 주세요.")
+                    println("id를 입력해 주세요.") // 삭제할 id 입력
                     continue
                 }
                 val article = articleRepository.getArticleByID(id)
@@ -33,6 +117,7 @@ fun main() {
                     println("${id}번 게시물은 존재하지 않습니다.")
                     continue
                 }
+
                 println("번호 : ${article.id}")
                 println("작성날짜 : ${article.regDate}")
                 println("갱신날짜 : ${article.updateDate}")
@@ -41,6 +126,11 @@ fun main() {
             } // end of "/article/detail"
 
             "/article/modify" -> {
+                if (loginedMember == null) {
+                    println("로그인해주세요.")
+                    continue
+                }
+
                 val id = rq.getIntParam("id", 0)
 
                 if (id == 0){
@@ -53,6 +143,13 @@ fun main() {
                     println("${id}번 게시물은 존재하지 않습니다.")
                     continue
                 }
+
+                // 현재 게시물의 사용자와 로그인 사용자가 일치하는지 확인
+                if (article.id != loginedMember.id) {
+                    println("권한이 없습니다")
+                    continue
+                }
+
                 print("${id}번 게시물 새 제목 : ")
                 val title = readLineTrim()
                 print("${id}번 게시물 새 제목 : ")
@@ -68,13 +165,21 @@ fun main() {
 
                 val filteredArticles = articleRepository.getFilteredArticles(searchKeyword, page,10)
 
-                println("번호 / 작성날짜 / 갱신날짜 / 제목 / 내용")
+                println("번호 /      작성날짜      / 작성자 / 제목 / 내용")
                 for (article in filteredArticles) {
-                    println("${article.id} / ${article.regDate} / ${article.updateDate} / ${article.title} / ${article.body}")
+                    val writer = memberRepository.getMemberById(article.memberId)!!
+                    val writerName = writer.nickname
+
+                    println(" ${article.id} / ${article.regDate} / $writerName / ${article.title} / ${article.body}")
                 }
             } // end of "/article/list"
 
             "/article/delete" -> {
+                if (loginedMember == null) {
+                    println("로그인해주세요.")
+                    continue
+                }
+
                 val id = rq.getIntParam("id", 0)
 
                 if (id == 0) {
@@ -87,18 +192,78 @@ fun main() {
                     println("${id}번 게시물은 존재하지 않습니다.")
                     continue
                 }
+
+                // 현재 게시물의 사용자와 로그인 사용자가 일치하는지 확인
+                if (article.id != loginedMember.id) {
+                    println("권한이 없습니다")
+                    continue
+                }
+
                 articleRepository.deleteArticle(article)
+
+                println("${id}번 게시물을 삭제하였습니다.")
             } // end of "/article/delete"
+        } // end fo when
+    } // end of while
+
+    println("== SIMPLE SSG 끝 ==")
+} //end of Main
+
+data class Member(
+    val id: Int, val regDate: String, var updateDate: String, val loginId: String,
+    val loginPw: String, val name: String, val nickname: String, val cellphoneNo: String, val email: String)
+
+object memberRepository {
+    private val members = mutableListOf<Member>()
+    private var lastId = 0
+
+    fun join(loginId: String, loginPw: String, name: String, nickName: String, cellphoneNo: String, email: String): Int {
+        val id = ++lastId // 가입된 멤버 순서 id, loginId(user1)과 표기는 다르지만 일치 됨.
+        val regDate = Util.getNowDateStr()
+        val updateDate = Util.getNowDateStr()
+        members.add(Member(id, regDate, updateDate, loginId, loginPw, name, nickName, cellphoneNo, email))
+        //println(members)
+
+        return id
+    }
+
+    fun makeTestMembers() { // user1부터 user9까지 사용자 입력, user10은 존재하지 않음, user join하여 입력가능
+        for (id in 1..9) {
+            join("user${id}", "user${id}", "홍길동${id}", "사용자${id}",
+                "0101234123${id}", "user${id}@test.com")
         }
     }
 
-    println("== SIMPLE SSG 끝 ==")
+    // 파라메터로 입력 받은 loginId가, 즉 멤버가 존재하는지 확인
+    fun isJoinableLoginId(loginId: String): Boolean {
+        val member = getMemberByLoginedId(loginId)
+
+        return member == null
+    }
+
+    // 현재 저장되어있는 사용자와 로그인하려는 사용자 일치 비교
+    fun getMemberByLoginedId(loginId: String): Member? {
+        for (member in members) {
+            if (member.loginId == loginId) {
+                return member
+            }
+        }
+        return  null
+    }
+
+    fun getMemberById(memberId: Int): Member? {
+        for (member in members) {
+            if (member.id == memberId) {
+                return member
+            }
+        }
+        return null
+    }
 }
 
 data class Article(
-    val id: Int, val regDate: String, var updateDate: String, var title: String, var body: String,
+    val id: Int, val regDate: String, var updateDate: String,  val memberId: Int, var title: String, var body: String,
 )
-
 
 object articleRepository {
     val articles = mutableListOf<Article>()
@@ -113,17 +278,19 @@ object articleRepository {
         return  null
     }
 
-    fun addArticle(title: String, body: String) {
+    fun addArticle(memberId: Int, title: String, body: String): Int {
         val id = ++lastId
         val regDate = Util.getNowDateStr()
         val updateDate = Util.getNowDateStr()
 
-        articles.add(Article(id, regDate, updateDate, title, body))
+        articles.add(Article(id, regDate, updateDate, memberId, title, body))
+
+        return id
     }
 
     fun makeTestArticle() {
-        for (id in 1..100) {
-            addArticle("제목_$id", "내용_$id")
+        for (id in 1..20) {
+            addArticle(id % 9 +1, "제목_$id", "내용_$id") //1 ~ 10(멤버 수)
         }
     }
 
@@ -185,7 +352,7 @@ object articleRepository {
         return filteredArticles
     }
 
-} // end of articleRepository
+}// end of articleRepository
 
 class Rq(command: String) {
     // 전체 URL : /artile/detail?id=1
@@ -240,9 +407,9 @@ class Rq(command: String) {
                 }
             } // end of for(), 저장된 최종값을 불변 맵에 저장
             paramMapTemp.toMap()
-        } // end of paramMap
+        }// end of paramMap
         //println(paramMap)
-    } //end of init
+    }//end of init
 
 
     // 입력 : /article/detail?id=2&title=제목1
@@ -284,7 +451,7 @@ class Rq(command: String) {
         }
     }
 
-}
+}// end of Rq()
 
 object Util {
     fun getNowDateStr(): String {
